+++ packages/cozy-client/src/index.js
@@ -20,3 +20,4 @@ export {
   HasManyInPlace
 } from './associations'
 export { dehydrate } from './helpers'
+export { getQueryFromState } from './store'